morgan used for logging health router moved to new file
