+++ udagram-api/user-api/package.json
@@ -24,6 +24,7 @@
     "email-validator": "^2.0.4",
     "express": "^4.16.4",
     "jsonwebtoken": "^8.5.1",
+    "morgan": "^1.10.0",
     "pg": "^7.9.0",
     "reflect-metadata": "^0.1.13",
     "sequelize": "^5.21.4",
@@ -34,6 +35,7 @@
     "@types/bluebird": "^3.5.26",
     "@types/cors": "^2.8.6",
     "@types/express": "^4.16.1",
+    "@types/morgan": "^1.9.2",
     "@types/node": "^11.11.6",
     "@types/sequelize": "^4.27.44",
     "@types/validator": "^10.9.0",
